tests: refactor
